services/inventory-service/requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.29.0
sqlalchemy>=2.0.29
asyncpg>=0.29.0
pydantic>=2.7.0
pydantic-settings>=2.2.0

pika>=1.3.2
psycopg2-binary>=2.9.9

python-jose[cryptography]>=3.3.0
python-multipart>=0.0.9